Move repository creation
