+++ stripe-crypto-onramp/src/main/java/com/stripe/android/crypto/onramp/OnrampCoordinator.kt
@@ -7,11 +7,15 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelStoreOwner
+import com.stripe.android.Stripe
+import com.stripe.android.core.networking.DefaultStripeNetworkClient
+import com.stripe.android.core.version.StripeSdkVersion
 import com.stripe.android.crypto.onramp.di.DaggerOnrampComponent
 import com.stripe.android.crypto.onramp.di.OnrampComponent
 import com.stripe.android.crypto.onramp.model.LinkUserInfo
 import com.stripe.android.crypto.onramp.model.OnrampCallbacks
 import com.stripe.android.crypto.onramp.model.OnrampConfiguration
+import com.stripe.android.crypto.onramp.repositories.CryptoApiRepository
 import com.stripe.android.crypto.onramp.viewmodels.OnrampCoordinatorViewModel
 import com.stripe.android.link.LinkController
 import javax.inject.Inject
@@ -134,10 +138,20 @@ class OnrampCoordinator @Inject internal constructor(
                 )
             }
 
+            val cryptoApiRepository = CryptoApiRepository(
+                stripeNetworkClient = DefaultStripeNetworkClient(),
+                publishableKeyProvider = { "" },
+                stripeAccountIdProvider = { "" },
+                apiVersion = Stripe.API_VERSION,
+                sdkVersion = StripeSdkVersion.VERSION,
+                appInfo = Stripe.appInfo
+            )
+
             val viewModel = ViewModelProvider(
                 owner = viewModelStoreOwner,
                 factory = OnrampCoordinatorViewModel.Factory(
                     linkController = linkController,
+                    cryptoApiRepository = cryptoApiRepository,
                     onrampCallbacks = onrampCallbacks
                 )
             ).get(
